Changed "KillHopperStateCommand" to "StopHopperStateCommand" and changed "StopHopperCommand" to "ReverseHopperCommand". Updated code accordingly
