+++ RobotCode2020New/src/main/java/frc/robot/commands/hoppercommands/ReverseHopperCommand.java
@@ -10,38 +10,36 @@
 package frc.robot.commands.hoppercommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import edu.wpi.first.wpilibj2.command.button.JoystickButton;
+import frc.robot.subsystems.HopperSubsystem;
 
 //-------- COMMAND CLASS --------\\
 
-public class KillHopperStateCommand extends CommandBase {
+public class ReverseHopperCommand extends CommandBase {
 
-    // -------- DECLARATIONS --------\\
-    private boolean state;
+    //-------- DECLARATIONS --------\\
 
-    // private
-    // -------- CONSTRUCTOR --------\\
+    private HopperSubsystem m_HopperSubsystem;
+    private JoystickButton home;
+    
+    //-------- CONSTRUCTOR --------\\
 
-    public KillHopperStateCommand() {
-        state = false;
-    }
+    public ReverseHopperCommand(HopperSubsystem hopperSubsystem, JoystickButton home) {
+        m_HopperSubsystem = hopperSubsystem;
+        this.home = home;
+    }   
 
-    //-------- METHODS --------\\    
+    //-------- METHODS --------\\
 
     // Called when the command is initially scheduled.
     @Override
     public void initialize() {
-        state = !state;
     }
 
     // Called every time the scheduler runs while the command is scheduled.
     @Override
-    public void execute() {     
-
-    }
-
-    public boolean getState()
-    {
-        return state;
+    public void execute() {   
+        m_HopperSubsystem.setSpeed(-0.35);
     }
 
     // Called once the command ends or is interrupted.
@@ -52,6 +50,6 @@ public class KillHopperStateCommand extends CommandBase {
     // Returns true when the command should end.
     @Override
     public boolean isFinished() {
-        return false;
+        return !home.get();
     }
-} //end of class DefaultHopperCommand
+} // end of class ReverseHopperCommand
